requirements.txt
annotated-types==0.7.0
anyio==4.6.0
apturl==0.5.2
asgiref==3.8.1
attrs==24.2.0
autosub==0.3.12
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.6
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-background-tasks==1.2.8
duplicity==0.8.21
elevenlabs==1.9.0
exceptiongroup==1.2.2
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.15.4
fonttools==4.53.1
future==0.18.2
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.1.post1
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
progressbar==2.5
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pysrt==1.1.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-sourcemaps==1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
SecretStorage==3.3.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.4
sqlparse==0.5.1
srt==3.5.3
starlette==0.38.5
systemd-python==234
Telethon==1.36.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.2
usb-creator==0.3.7
virtualenv==20.26.3
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
